init/requirements.txt
cassandra-driver==3.29.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
geni-lib==0.9.9.4
geomet==0.2.1.post1
idna==3.10
ipaddress==1.0.23
lxml==5.3.0
pycparser==2.22
requests==2.32.3
six==1.17.0
urllib3==2.2.3
wrapt==1.17.0
